Bugfix: Player now can't destroy planet that is original in test mode
Before the error was logged out because of that we were entering in basic editor, but we didn't assign new Current Planet, so we wanted to get infos of null. Now we are making sure that when we can't destroy planet we want to get that planet as our Current Planet in editors.