WIP: Logic Levels - a bit of bug fixing and prototyping
Fixed bug of controlling logic areas - EventSystem.IsOverGameObject is detecting gameobjects that are in canvas - not all, so it didn't detect my panel which was normal Sprite2D.
Also started prototyping ActionArea - Area for player's limitation. Maker of level will be able to design level in such way, where player is able to put planets only in some zones - areas.